Move CodeInfo, InlinedFd, and Sym definitions into symbolize/mod.rs
The CodeInfo, InlinedFd, and Sym type definitions are somewhat hard to
find inside symbolize/symbolizer.rs and also conceptually don't exactly
belong in there.
With this change we move them into symbolize/mod.rs.
